+++ trains_front/src/pages/RoutesDisplay.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+import '../styles/RoutesDisplay.css';
+import '../styles/Login.css'
+import '../styles/Main.css'
+import tom from '../images/sad_tom.png';
+import { useLocation } from 'react-router-dom';
+import { Link } from "react-router-dom"
+import { Card, Button } from 'react-bootstrap';
+import Form from "../components/Form";
+import RouteCards from "../components/RouteCards"
+
+export default function RoutesDisplay() {
+
+    return (
+        <>
+            <div className='cardWrapper'>
+                <div className="home--link"><Link to="/"><span className="black">TRAIN</span> 
+                    <span className="blue">SERVICE</span></Link></div>
+                <Form />
+                <RouteCards/>
+
+            </div>
+        </>
+    );
+}
